+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Display all comments when no finding/section is selected
 * Add information about superuser privileges when creating API tokens
 * Migrate documentation from mkdocs to vitepress
+* Support multi select and bulk actions in report sidebar
+* Highlight inactive users in project member selection
 
 
 ## v2026.36 - 2026-05-20